// ExistingCode/src/data-access/domain/tables/collarcoordinate/schema.ts
export const CollarCoordinateTableInsertBaseSchema = z.object({
	/**
	 * dataType: uniqueidentifier
	 * defaultValue: (newid())
	 */
	collarCoordinateId: z.any().optional(),
	/**
	 * dataType: nvarchar
	 */
	organization: z.string().max(30),
	/**
	 * dataType: uniqueidentifier
	 */
	collarId: z.any(),
	/**
	 * dataType: smallint
	 * defaultValue: ((0))
	 */
	priority: z.number().int().optional(),
	/**
	 * dataType: nvarchar
	 * defaultValue: (N'Archived')
	 */
	priorityStatus: z.string().max(50).optional(),
	/**
	 * dataType: nvarchar
	 */
	grid: z.string().max(50).nullable().optional(),
	/**
	 * dataType: decimal
	 */
	east: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	north: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	rL: z.number().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	rLSource: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	surveyMethod: z.string().max(50),
	/**
	 * dataType: datetimeoffset
	 */
	surveyOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	surveyBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	surveyCompany: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	instrument: z.string().max(50).nullable().optional(),
	/**
	 * dataType: bit
	 */
	isDeleted: z.any().nullable().optional(),
	/**
	 * dataType: bit
	 */
	validated: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	comments: z.string().max(1000).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	dataSource: z.string().max(255),
	/**
	 * dataType: int
	 */
	migrationID: z.number().int().nullable().optional(),
	/**
	 * dataType: binary
	 */
	migrationRv: z.any().nullable().optional(),
	/**
	 * dataType: geography
	 */
	geoPointWGS: z.any().nullable().optional(),
	/**
	 * dataType: geometry
	 */
	geoPoint: z.any().nullable().optional(),
	/**
	 * dataType: tinyint
	 */
	validatedStatus: z.number().int(),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	reportIncludeInd: z.any().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	validationStatus: z.number().int().optional(),
	/**
	 * dataType: nvarchar
	 */
	validationErrors: z.string().max(4000).nullable().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	rowStatus: z.number().int().optional(),
	/**
	 * dataType: uniqueidentifier
	 */
	supersededById: z.any().nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((1))
	 */
	activeInd: z.any().optional(),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	createdOnDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 * defaultValue: (CONVERT([nvarchar](50),coalesce(session_context(N'UserName'),suser_sname())))
	 */
	createdBy: z.string().max(50).optional(),
	/**
	 * dataType: datetimeoffset
	 */
	modifiedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	modifiedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: timestamp
	 */
	rv: z.date(),
});